requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.3
MarkupSafe==2.1.2
flask-sqlalchemy==3.0.3
pillow==10.4.0
torch==2.6.0
torchvision==0.21.0
numpy>=1.26.0
pandas>=2.0.0
python-dotenv==0.19.0
gunicorn==20.1.0
requests>=2.31.0
groq>=0.4.0
fluvio-python>=0.1.0
websockets>=10.0
flask-socketio==5.3.6 